AEM-3796 - Testimonials threshold added to swipe, moved to src and used ui-events.
